requirements.txt
streamlit>=1.31
pandas>=2.0
openpyxl>=3.1
python-dotenv>=1.0
SQLAlchemy>=2.0
psycopg2-binary>=2.9
PyYAML>=6.0

streamlit-aggrid
streamlit-echarts
streamlit-extras
streamlit-shadcn-ui>=0.1.19
playwright>=1.53
streamlit-antd-components>=0.3.2
